+++ src/app/(auth)/signup/page.tsx
@@ -2,8 +2,6 @@
 
 import { SignupTemplate } from "@/app/features/signup/ui/template/SignupTemplate";
 
-
-
 export default function page() {
   return (
     <>
