+++ src/app/features/show/api/fetchGroupDetail.ts
